Separated statistics classes into individual py files and own folder
